64-bit fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@1057 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/MyDetails/data.cpp
@@ -123,7 +123,7 @@ Protocol::~Protocol()
 {
 }
 
-void Protocol::lcopystr(TCHAR *dest, TCHAR *src, int maxlen)
+void Protocol::lcopystr(TCHAR *dest, TCHAR *src, size_t maxlen)
 {
 	if (lstrcmp(dest, src) != 0)
 	{
